Created ConnectivityService and implemented it in baseViewModel
