+++ IntelliHouse2000App/Services/Connectivity/IConnectivityService.cs
@@ -1,0 +1,8 @@
+﻿using IntelliHouse2000App.ViewModels;
+
+namespace IntelliHouse2000App.Services.Connectivity;
+
+public interface IConnectivityService
+{
+    void Init(BaseViewModel viewModel);
+}
